Moved storage helper functions
