+++ store/toml/redirect.go
@@ -1,18 +1,10 @@
 package toml
 
 import (
-	"bytes"
-	"github.com/BurntSushi/toml"
 	"github.com/satori/go.uuid"
 	"github.com/tboerger/redirects/model"
 	"github.com/tboerger/redirects/store"
-	"io/ioutil"
 )
-
-// redirectCollection represents the internal storage collection.
-type redirectCollection struct {
-	Redirects []*model.Redirect `toml:"redirects"`
-}
 
 // GetRedirects retrieves all redirects from the TOML store.
 func (db *data) GetRedirects() ([]*model.Redirect, error) {
@@ -113,38 +105,4 @@ func (db *data) CreateRedirect(create *model.Redirect) error {
 	)
 
 	return db.write(root)
-}
-
-// load parses all available records from the storage.
-func (db *data) load() (*redirectCollection, error) {
-	res := &redirectCollection{
-		Redirects: make([]*model.Redirect, 0),
-	}
-
-	content, err := ioutil.ReadFile(db.file)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if _, err := toml.Decode(string(content), res); err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
-
-// write writes the TOML content back to the storage.
-func (db *data) write(content *redirectCollection) error {
-	buf := new(bytes.Buffer)
-
-	if err := toml.NewEncoder(buf).Encode(content); err != nil {
-		return err
-	}
-
-	if err := ioutil.WriteFile(db.file, buf.Bytes(), 0640); err != nil {
-		return err
-	}
-
-	return nil
 }
