+++ store/toml/toml.go
@@ -1,8 +1,11 @@
 package toml
 
 import (
+	"bytes"
 	"fmt"
+	"github.com/BurntSushi/toml"
 	"github.com/tboerger/redirects/config"
+	"github.com/tboerger/redirects/model"
 	"github.com/tboerger/redirects/store"
 	"io/ioutil"
 	"os"
@@ -10,6 +13,11 @@ import (
 	"strings"
 	"sync"
 )
+
+// collection represents the internal storage collection.
+type collection struct {
+	Redirects []*model.Redirect `toml:"redirects"`
+}
 
 // data is a basic struct that iplements the Store interface.
 type data struct {
@@ -25,6 +33,40 @@ func (db *data) Name() string {
 // Config just returns a simple configuration explanation.
 func (db *data) Config() string {
 	return fmt.Sprintf("file:%s", db.file)
+}
+
+// load parses all available records from the storage.
+func (db *data) load() (*collection, error) {
+	res := &collection{
+		Redirects: make([]*model.Redirect, 0),
+	}
+
+	content, err := ioutil.ReadFile(db.file)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if _, err := toml.Decode(string(content), res); err != nil {
+		return nil, err
+	}
+
+	return res, nil
+}
+
+// write writes the TOML content back to the storage.
+func (db *data) write(content *collection) error {
+	buf := new(bytes.Buffer)
+
+	if err := toml.NewEncoder(buf).Encode(content); err != nil {
+		return err
+	}
+
+	if err := ioutil.WriteFile(db.file, buf.Bytes(), 0640); err != nil {
+		return err
+	}
+
+	return nil
 }
 
 // New initializes a new TOML store.
